pyproject.toml
[build-system]
requires = ["flit_core"]
build-backend = "flit_core.buildapi"

[project]
name = "mdformat-configurable-black"
authors = [
    { name = "veghdev", email = "david.andras.vegh@gmail.com" },
]
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["mdformat", "markdown", "markdown-it"]
requires-python = ">=3.7"
dependencies = [
    "mdformat",
    "markdown-it-py",
    "black",
]
version="0.0.1"
description="mdformat-configurable-black plugin"

[project.urls]
Homepage = "https://github.com/veghdev/mdformat-configurable-black"

[project.entry-points."mdformat.parser_extension"]
configblack = "mdformat_configurable_black:ConfigurableBlack"

[tool.flit.sdist]
include = []
exclude = [".github/"]
